refactor(dl): reorganize package structure
- Move ContentDisposition to pkg/attachment as shared package
- Move service implementations (oci, ks3) to internal/service/
- Clean up dl root to only contain config files
