+++ dl/pkg/attachment/content_disposition.go
@@ -1,4 +1,4 @@
-package dl
+package attachment
 
 import (
 	"fmt"
