+++ dl/pkg/attachment/content_disposition_test.go
@@ -1,4 +1,4 @@
-package dl
+package attachment
 
 import (
 	"net/http"
